% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdata.R, R/data.R
\docType{data}
\name{b}
\alias{b}
\alias{b}
\title{Example data from HW 1, problem 3}
\format{An object of class \code{numeric} of length 30.}
\usage{
b

b
}
\description{
Example data from HW 1, problem 3

Example data from HW 1, problem 3
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p3.txt}

\url{http://www.stat.umn.edu/geyer/3701/data/q1p3.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}

Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
