% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsmyplot.R
\name{filterFun}
\alias{filterFun}
\title{Wrapper function for dplyr}
\usage{
filterFun(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
dplyr
}
\description{
Wrapper function for dplyr
}
\examples{
data(d)
dplyr::filter(d, p>0.05)

}
