% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{Fun1}
\alias{Fun1}
\title{Problem 1
matrix multiplication with user checks}
\usage{
Fun1(A, x)
}
\arguments{
\item{x}{vector}

\item{a}{matrix}
}
\value{
vector (1 dimensional)
}
\description{
multiplies transposed vector by the inverse of a matrix times the vector
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
Fun1(a, x)

}
