% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{Fun3}
\alias{Fun3}
\title{same function for problem 3 and problem 4
standardizes columns of numeric matrix}
\usage{
Fun3(A)
}
\arguments{
\item{a}{matrix}
}
\value{
matrix
}
\description{
subtracts the mean of a matrix from the matrix and divides by the matrix's standard deviation
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p3.rda"))
Fun3(a)

}
