% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{func1}
\alias{func1}
\title{Problem 1
Calculate Mean, Variane, SD}
\usage{
func1(x)
}
\arguments{
\item{g}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and sd of a vector
}
\examples{
func1(rnorm(10))
}
