% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{func2}
\alias{func2}
\title{Problem 2
Calculate Mean, Variane, SD (again)}
\usage{
func2(x)
}
\arguments{
\item{g}{vector}
}
\value{
list
}
\description{
Computes the mean, variance and sd of a vector, but with user checks
}
\examples{
func2(rnorm(10))
}
