% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{func3}
\alias{func3}
\title{Problem 3
MLE of gamma distribution}
\usage{
func3(x)
}
\arguments{
\item{b}{vector}
}
\value{
scalar
}
\description{
Computes the liklihood of a gamma distribution
}
\examples{
func3(rnorm(10))
}
