% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdata.R
\docType{data}
\name{g}
\alias{g}
\title{Example data from HW 1, problem 1}
\format{An object of class \code{numeric} of length 30.}
\usage{
g
}
\description{
Example data from HW 1, problem 1
}
\references{
\url{http://www.stat.umn.edu/geyer/3701/data/q1p1.txt}
}
\author{
Charlie Geyer \email{geyer@umn.edu}
}
\keyword{datasets}
