% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{\%Fun2\%}
\alias{\%Fun2\%}
\title{Problem 2
matrix multiplication with user checks using a binary operator}
\usage{
A \%Fun2\% x
}
\arguments{
\item{x}{vector}

\item{a}{matrix}
}
\value{
vector (1 dimensional)
}
\description{
Problem 2
matrix multiplication with user checks using a binary operator
}
\examples{
load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
a \%Fun2\% x

load(url("http://www.stat.umn.edu/geyer/3701/data/q2p1.rda"))
}
