% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syedToolsRfunctions.R
\name{myapply}
\alias{myapply}
\title{Problem 5
myapply function}
\usage{
myapply(X, MARGIN, FUN, ...)
}
\arguments{
\item{x}{matrix}
}
\value{
varies
}
\description{
a function like the apply() function
}
\examples{
fred <- matrix(1:6, ncol = 2)
myapply(fred,1,"mean")
myapply(fred,2,"mean")

myapply(fred,1,"max")
myapply(fred,2,"max")

myapply(fred,1,function(x) quantile(x,0.75))
myapply(fred,2,function(x) quantile(x,0.75))

myapply(fred, 1, quantile, probs=.75)
myapply(fred, 2, quantile, probs=.75)

myapply(fred, 1, quantile, probs = (1:3)/4)
myapply(fred, 2, quantile, probs = (1:3)/4)

moo <- myapply(fred, 1, mean)
fred.mean.free.rows <- sweep(fred, 1, moo)
rowMeans(fred.mean.free.rows)
}
